[GH-50] Pandas 1.0 compatible issue. (#51)
Pandas 1.0 introduces a `_ensure_type` method that asserts the type of
`self` must be the same type as the input `obj`.  In this lib, the type
of `self` is `StockDataFrame` and the type of the object is `DataFrame`
which breaks this assert.

The temporary fix is to bypass the assert by override the `_ensure_type`
method in `StockDataFrame`.  The is just a work around.  The good news
is that the original method doesn't contain any business logic.

Other changes:
* Update the travis configuration to add CI for python 3.8.
* Drop the CI build for 3.4.  It's too old and takes too long to run.