[GH-122] Update calculation of `middle`
The typical price or middle price is calculated by:
```
(high + low + close) / 3
```

But it would be accurate if the data source has the `amount` that
represents the total amount of capital traded in that period.
Use `middle = amount / volume` when amout is available.

Also fix some typos in the read me file.
Add some utilities for dropping columns, head or tail.

Add python 3.11 in CI and drop 3.9.